renamed:    src/front/js/pages/info.js -> src/front/js/component/info.js 	modified:   src/front/js/layout.js 	new file:   src/front/js/pages/infoProducto.js
